convert cookie-notice to trait
